+++ Plugins/StreamChat/Source/StreamChat/Private/StreamChatClientComponent.cpp
@@ -393,7 +393,7 @@ void UStreamChatClientComponent::QueryChannel(const FChannelProperties& ChannelP
             }
             UChatChannel* Channel = WeakThis->CreateChannelObject(Dto);
 
-            WeakThis->Channels.Add(Channel);
+            WeakThis->Channels.Insert(Channel, 0);
             WeakThis->ChannelsUpdated.Broadcast(WeakThis->Channels);
             if (Callback)
             {
