+++ Plugins/StreamChat/Source/StreamChatUi/Public/Common/SStreamListView.h
